+++ weni/internal/urls.py
@@ -17,6 +17,8 @@ from weni.internal.tickets.urls import urlpatterns as tickets_urls
 from weni.internal.classifier.urls import urlpatterns as classifier_urls
 from weni.internal.channel.urls import urlpatterns as channel_urls
 from weni.internal.statistic.urls import urlpatterns as statistics_urls
+from weni.internal.globals.urls import urlpatterns as globals_urls
+from weni.internal.externals.urls import urlpatterns as externals_urls
 
 
 internal_urlpatterns = []
@@ -27,6 +29,8 @@ internal_urlpatterns += tickets_urls
 internal_urlpatterns += classifier_urls
 internal_urlpatterns += channel_urls
 internal_urlpatterns += statistics_urls
+internal_urlpatterns += globals_urls
+internal_urlpatterns += externals_urls
 
 
 urlpatterns = [path("internals/", include(internal_urlpatterns))]